requirements.txt
cbor2
roboticstoolbox-python==1.1.1
numpy==1.26.4
swift-sim
coppeliasim-zmqremoteapi-client
scipy
matplotlib